Updated README.md
Resolving the issue of unmatched VirtualBox Guest Additions requires
the installation of vagrant-vbguest plugin prior to setup and no
further changes to configuration. Hence, the README.md is updated with
a small section with regards to vagrant-vbguest.

Minor changes to the README.md was also made to enhance the
instructions.